frontend/pyproject.toml
[tool.poetry]
name = "plouf-frontend"
version = "0.1.0"
description = "User interface for Plouf appllication."
authors = ["tristanqtn <tristan.querton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
plotly = "^5.24.1"
numpy = "^2.2.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit = "^1.42.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
